+++ authpass/lib/l10n/app_localizations_et.dart
@@ -146,6 +146,22 @@ class AppLocalizationsEt extends AppLocalizations {
   }
 
   @override
+  String get unableToLaunchUrlTitle => 'Unable to open Url';
+
+  @override
+  String unableToLaunchUrlDescription(Object url, Object openError) {
+    return 'Unable to launch ${url}: ${openError}';
+  }
+
+  @override
+  String get unableToLaunchUrlNoHandler => 'No application available for url.';
+
+  @override
+  String launchedUrl(Object url) {
+    return 'Opened URL: ${url}';
+  }
+
+  @override
   String get menuItemGeneratePassword => 'Generate Password';
 
   @override
@@ -171,6 +187,9 @@ class AppLocalizationsEt extends AppLocalizations {
 
   @override
   String get menuItemAbout => 'About';
+
+  @override
+  String get actionOpenUrl => 'Open URL';
 
   @override
   String get passwordPlainText => 'Reveal password';