+++ authpass/CHANGELOG.md
@@ -1,5 +1,6 @@
 # (unreleased)
 
+* Allow opening of URLs of entries. (On desktop using shift+cmd+U). #140
 * Disable autocorrect/suggestions and text capitalization for URL and Username fields. #145
 * Translations: Added Indonesian.
 * DEV: update flutter to 1.22.0-12.1.pre (from 1.22.0-1.0.pre)